📝 Move data grid example to a separate menu item
Hopefully this will clarify that this is just for testing, and not for real use...
